[Hongtao] clean up useless code
